Model description

Purpose and pattern

This model has two purposes, the first is to build a stable model for BPH population. The second is to research the effectiveness of cultivated flower in preventing the dispersion of BPH and providing a rough guideline on how flowers should be planted to effectively prevent the BPH from spreading and still, ensure economic benefit for the farmers.

To verify the stability and trustiness of the model, we use population structure and patterns from Syahrawati2019 as a pattern for verifying our model.

Entities, state variables, and scales

Entities and state variables

Our model contains two types of entities.

The first type of entity is the BPH. Each is associated with, energy, a stage, gender, and type. The BPH has a truncate-winged form Brachypterous and a fully-winged form Macropterous. The Brachypterous variant moves slowly and has a high reproduction rate, while the other has the opposite specifications. We also employ a countdown-based system to model the reproduction and growth of BPHs.

Variable	Description	Value(s)

$i$	The agent identifier number	$1, 2, \ldots$

$x_i$	The $x$-coordinate of agent $i$ on the environment grid	$1, 2, \ldots S$

$y_i$	The $y$-coordinate of agent $i$ on the environment grid	$1, 2, \ldots S$

$z^{(g)}_i$	The gender of agent $i$ (male, female)	$0, 1$

$z^{(s)}_i$	The current stage of agent $i$ (egg, nymph, adult, dead)	$0, 1, 2, 3$

$z^{(f)}_i$	The form of agent $i$ (brachypterous, macropterous)	$0, 1$

$t^{(s)}_i$	The countdown to the next stage of agent $i$	$0, 1,\ldots$

$t^{(r)}_i$	The countdown to the next reproduction of agent $i$	$0, 1,\ldots$

The second type of entity is the environment (grid) cell. An environment cell can plant rice or cultivated flower. Rice cells have living energy while flower cells do not. But for the sake of simplicity, we model them in one entity and have the flower cells ignore the energy. All cells have an elimination probability, which is the probability that the BPH on that cell will be eliminated. The cultivated flowers do not directly eliminate the BPH, they attract their natural enemies (such as bees). The natural enemies wander around, therefore every cell has an elimination probability.

Info

Since we use Julia's matrices to represent the simulated environment, there is no such thing as a GridCell type with all the described state variable in the implementation. These state variables are written is this way for the sake of model description.

Variable	Description	Value(s)

$x$	The cell's $x$-coordinate on the environment grid	$1, 2,\ldots,S$

$y$	The cell's $y$-coordinate on the environment grid	$1, 2,\ldots,S$

$t_{x, y}$	The type of the cell at $x, y$ (flower, rice)	$0, 1$

$e_{x, y}$	The energy of the cell at $x, y$	$[0, 1]$

$p_{x, y}$	The elimination probability of the cell at $x, y$	$[0, P_0]$

Scales

Regarding the scales of the model, we use a discrete space scale. The total simulated environment is equivalent to a $360\mathrm{m^2}$ rice field. The field is divided into $125\times125$ grid cells; i.e. each cell's width is roughly equivalent to $0.15\mathrm{m}$ real world length.

Time is also represented discretely in the model. Each time step is equivalent to one actual hour. The total duration of the simulation is $2880$ steps, or $120$ days, equivalent to a crop cycle.

Process overview and scheduling

Scheduling

In each step of the simulation, the BPH agents execute their processes and then the environment cells do theirs. The BPH agents with higher energy perform their actions first.

The vegetation cells process

The environment only has one action, which is to eliminate the BPHs. This action does not apply if there is no flower planted.

The BPH agents processes

The agent actions are determined by their stages, each stage have different behaviours.

RiceBPH.Models.agent_step_egg! — Function

agent_step_egg!(agent, model)

The eggs have the following actions (performed in order):

agent_action_growup!

agent_action_die!

RiceBPH.Models.agent_step_nymph! — Function

agent_step_nymph!(agent, model)

The nymphs have the following actions (performed in order):

agent_action_growup!

agent_action_move!

agent_action_eat!

agent_action_die!

RiceBPH.Models.agent_step_adult! — Function

agent_step_adult!(agent, model)

The adult BPHs have the following actions (performed in order):

agent_action_growup!

agent_action_move!

agent_action_eat!

agent_action_reproduce!

agent_action_die!

Design Concepts

Basic principle

The model is based on a natural real-life phenomenon called "hopper burn", where a small infestation of BPH spreads out and destroys the whole rice field.

We model bio-conversion using energy. The energy is used to measure how well an agent is living. The energy accepts values ranging from zero to one. When an agent's energy goes to zero, it dies and gets removed from the simulation. Both entities have their energy scaled to $0$ and $1$ and there is a conversion parameter to represent the relative energy scale between them.

The flower grid cells are simulations of real-life cultivating flowers. The flowers invite the BPH's natural enemies to eliminate them. We emulate this with an elimination probability map on the whole field. Each cell has a probability that: if a BPH is in that position, it might get eliminated.

Emergence

At a high level, we observe the BPH population structure and the overall healthiness of the rice field.

Adaptation

The BPHs base their decisions on their internal states and on the vegetation cell's type and energy.

The first input is their stage. Each stage links to specific actions. For example, BPH eggs can not move around. The second input is their countdowns. Their actions are timed by their countdowns. The third input is their energy. Their actions have energy-based conditions. The fourth input is the environment cell type and energy. These factors affect their movement, eating, and reproduction.

Prediction

The BPHs agent implicitly predicts the amount of food available. Their prediction is presented in probabilities. Their movement distribution is calculated from environment energy states. Their reproduction is also conditioned on the environment cell's energy.

Sensing

The BPHs are able to sense their internal states. They can also sense the energy and type of nearby cells to make predictions and decisions.

Interaction

There are one direct interaction and two indirect ones.

The direct one is that the BPHs feed off the cells with rice directly. The indirect interactions are:

The cells eliminate the BPHs with the probability map.

The BPHs indirectly compete among themselves for food from the vegetation cells.

We consider the former interaction indirect because the probability map was just emulating the natural enemies, which is not modeled in our model.

Stochasticity

Stochasticity is used in the model mainly for the diversity of the results and to realistically model multi-option processes and to simplify complicated phenomena.

The processes that use stochasticity are:

initial placement and population structure of BPHs,

the BPHs' growth,

the BPHs' reproduction,

the BPHs' movement.

For reproducibility, we use a controlled pseudo-random number generator for all the stochastic processes in the model.

Observation

During our experiment, we collect the following data:

the percentage of healthy rice cells $r_R$,

the number of eggs $n_E$,

the number of nymphs $n_N$,

the number of adult macropterous $n_M$,

the number of adult brachypterous $n_P$,

the number of female nymphs and adults $n_F$.

For the BPH-related data, we also calculate inferable data such as:

the total number of BPHs (eggs are not counted) $n_A$,

the percentage of nymphs $r_N$,

the percentage of adult macropterous $r_M$,

the percentage of adult brachypterous $r_P$,

the ratio of females over males $r_F$.

We collect these data at each time step of the simulation. We use a function-like annotation to refer to the data point at a specific time step. For example, the number of eggs at the time $t$ is $n_E(t)$.

Initialisation

The initialization of the simulation includes several steps. The first step is to calculate the model states and initialize vegetation cells. The model state also creates a controlled random number generator. The second step is to initialize the BPH agents and place them inside the model. At the initialization of the model, we create the environment.

Warning

TODO?

RiceBPH.Models.init_model — Function

init_model(; seed::Union{Int, Nothing} = Nothing, kwargs...)

Create and return a ABM model. Keyword arguments are passed to ModelParameters.

At the initialization of the model, inferable model properties are calculated to optimize the runtime of the simulation (see init_properties). This process also create the simulated environment and the vegetation cell state.

After that, the BPH agents are initialized with random states. After the BPHs are initialized and placed inside the model, we collect the initial data of interest and store them in the model state.

Input data

The model does not use any external data.

Submodels

Notation

To make sure readers understand our model correctly, we will provide the notation system that we use in this material.

We use index variables such as $i,j,k,l,\ldots$ to refer to the identification number of an agent. The state variable of an agent is referred to by their respective notation in Entities, state variables, and scales. For example, the energy of agent $i$ is $e_i$.

For vegetation cells, we use spatial coordinates (such as $x,y$) to index them. Their state variables follow the same rule. For example, the energy of the rice cell at position $x, y$ is $e_{x,y}$, the energy of the rice cell where the agent $i$ at is $e_{x_i, y_i}$.

To abuse the notation, we hide the timestep variable and consider each submodel within its context (the "current" timestep). The state variable is updated in-place and the update is notated by $\gets$. For example, when an agent consumes energy:

\[e_i \gets e_i - E_C.\]

We do not explicitly refer to the timestep $t$ for several reasons. The main reason is that sometime, state variables are updated many times within a timestep, so it is not possible to use a timestep variable to describe the changes. Moreover, every submodel is described in its own context, therefore, the current time step is always $t$; adding the time step variable $t$ does not make it easier to perceive the model (if not harder).

Model parameters

The initialisation variables, parameters and states of the model are presented in the next table:

Variable	Description	Value

$S$	The environment grid size	$125$

$S_F$	The number of flower cells at the center of the grid	$0$

$N_{I}$	The number of initialized BPH	$200$

$P_0$	The base elimination probability	$0$

-	The BPHs' initial positions	Corner

$E_{T}$	The energy conversion from rice to BPHs	$0.032$

$E_{C}$	The energy conversion from rice to BPHs	$E_T / 3$

Initialisation actions

Since the vegetation cells are organized in a grid, we use matrices to represent the cells for efficiency. The coordinates of the vegetation cells translate directly to the matrix indices. Each of the other states is represented by a matrix.

RiceBPH.Models.init_cell_types — Function

init_cell_types(map_size::Integer, flower_width::Integer)
init_cell_types(S::Integer, S_F::Integer)

Returns a CellType matrix. Rice cells' values are 1. Flower cells' values are 0. The resulting matrix can be defined by:

\[\begin{equation} \left[t_{x,y}\right]_{S \times S}=\left\{ \begin{array}{cl} 0, & y_{a}\le y\le y_{b},\\ 1, & \text{otherwise}, \end{array}\right\}, \end{equation}\]

where $y_a$ and $y_b$ are auxilary variables defined by:

\[\begin{align} y_{a} & =\left\lfloor \frac{S-S_{F}}{2}\right\rfloor +\left(S+S_{F}\mod2\right)+1.\\ y_{b} & =S+S_{F}-1. \end{align}\]

See also: CellType.

RiceBPH.Models.init_rice_map — Function

init_rice_map(S::Integer)

Returns a matrix filled with $1$ of size $S \times S$.

\[\left[e_{x,y}\right]_{S\times S}\equiv1.\]

In code terms, it's basically ones(Float32, S, S).

RiceBPH.Models.init_pr_eliminate — Function

init_pr_eliminate(init_pr::Float32, cell_types::Matrix{CellType})

Returns Float32 matrix of elimination probability. The result matrix is defined by:

\[\left[p_{x,y}\right]_{S\times S} = P_0 \times (1 - M)) ∗ G(\sigma),\]

where:

$G(\sigma)$ is the Gaussian kernel,

$P_0$ is the base elimination probability,

$M$ is the cell type matrix (see init_cell_types).

RiceBPH.Models.init_bphs! — Function

init_bphs!(model)

Create $N_I$ agents and put them in the model. The state variables of agent $i$ are initialized as follows:

positions $x_i, y_i$: sampled from cumulative distribution $\mathrm{Poi}(6.00)$ with constraint $x_i \le 15$ and $y_i \le 15$ (see IP_MAX and IP_DST),

energy $e_i$: sampler from $\mathcal{N}(0.50, 0.50)$,

gender $z^{(g)}_i$: sample from {Male, Female} with weight {0.37, 0.63},

form $z^{(f)}_i$: sample from {Brachy, Macro} with weight {0.47, 0.53},

stage $z^{(s)}_i$: sample from {Egg, Nymph, Adult, Dead} with weight {0.33, 0.47, 0.20, 0.00},

stage countdown $t^{(s)}_i$: derives from their other state variables,

reproduction countdown $t^{(r)}_i$: is the preoviposition countdown if agent is not an adult, else the countdown is either preoviposition countdown or reproduction countdown with equal chances.

Vegetation cell submodels

RiceBPH.Models.model_action_eliminate! — Function

model_action_eliminate!(model)

Eliminate BPHs base on their energy and the pr eliminate map. The elimination probability is given by $p_{x,y}$.

For each position $x, y$, for each agent $i$ whose position $x_i = x, y_i = y$, the agent is removed from the simulation with the probability $p_{x, y}$.

BPH agent submodels

RiceBPH.Models.agent_action_growup! — Function

agent_action_growup!(agent, model)

Perform the grow up action on agent with id $i$:

The stage countdown of the agent is decresed by one.

\[\begin{equation} t_i^{(s)} \gets t_i^{(s)} - 1. \end{equation}\]

The agent then consumes energy.

\[e_i \gets e_i - E_C\]

If stage countdown is greater than zero, end the action. Otherwise, get the next stage and sample the next stage countdown, assign the next stage and the next countdown to the agent. The next stage $z^{(s)}_i$ and next the stage countdown $t^{(s)}_i$ is returned by get_next_stage.

RiceBPH.Models.agent_action_move! — Function

agent_action_move!(agent, model)

Perform the move action of the agent:

Let the agent's ID $i$, the agent's position on the grid $x=x_i$, $y=y_i$. The agent only moves under certain conditions, which are: energy is greater or equal to energy consumption

\[\begin{equation} e_i \ge E_T; \end{equation}\]

or the cell that the agent is currently on is a rice cell and has zero energy:

\[\begin{equation} e_{x, y} = 0; \end{equation}\]

or the cell type is flower:

\[\begin{equation} t_{x, y} = 0. \end{equation}\]

If the conditions are not met, stop the action. Otherwise, the agent then consumes energy:

\[\begin{equation} e_{i} \gets e_{i} - E_C. \end{equation}\]

After that, a direction $\Delta x, \Delta y$ is sampled within the radius of 2 cells (approximately 30cm) with weights. The weight of the moving direction is calculated by

\[\begin{equation} w_{\Delta x,\Delta y}=\begin{cases} e_{x+\Delta x,y+\Delta y}, & t_{x+\Delta x,y+\Delta y}=1,\\ 0.5, & t_{x+\Delta x,y+\Delta y}=0. \end{cases} \end{equation}\]

After a direction is sampled, the agent then moves along the sampled direction

\[\begin{align} x_i &\gets x_i + \Delta x,\\ y_i &\gets y_i + \Delta y. \end{align}\]

RiceBPH.Models.agent_action_eat! — Function

agent_action_eat!(agent, model)

Perform eat action of agent.

First, check if the current cell is a FlowerCell, if it is, stop the action.

An amount of energy is subtracted from the RiceCell $x_i,y_i$ and added to the agent $i$.

\[\begin{align} e_{x_i, y_i} &\gets e_{x_i,y_i} - \Delta e,\\ e_i &\gets e_i + \Delta e, \end{align}\]

where the transfered energy $\Delta e$ is calculated by

\[\begin{equation} \Delta e=\min\left(1-e_i,e_{x_i,y_i},e_{T}\right). \end{equation}\]

RiceBPH.Models.agent_action_reproduce! — Function

agent_action_reproduce!(agent, model)

Perform the reproductive action of agent $i$.

First, decrease the reproduction countdown of the agent $i$.

\[\begin{equation} t^{(p)}_i \gets t^{(p)}_i - 1 \end{equation}\]

After that, check for the reproduction conditions: the agent is a female,

\[\begin{equation} z^{(g)}_i = 1, \end{equation}\]

the agent energy is larger or equals than the energy consumption

\[\begin{equation} e_i \ge E_C, \end{equation}\]

the reproduction countdown is zero:

\[t^{(p)} = 0,\]

the energy of rice cell at agent position is greater than the energy transfer parameter:

\[e_{x_i, y_i} \ge E_T.\]

If the condition is met, the number of offsprings $N$ from the distributions of offspring quantity $\mathcal{N}(8.50, 3.50)$.

For $k=\overline{1, N}$, the offspring agent $k$ is initialized with:

energy the same as their parent: $e_k \gets e_i$,

the position of the parent $x_k \gets x_i$, $y_k \gets y_i$,

stage is egg: $z^{(s)}_k \gets 0$,

stage countdown is of egg: $t^{(s)}_k \gets$$\mathcal{N}(228.00, 84.00)$,

gender $z^{(g)}_k$: sample from {Male, Female} with weight {0.37, 0.63},

form $z^{(f)}_k$: sample from {Brachy, Macro} with weight {0.47, 0.53},

reproduction countdown follows the preoviposition distribution (which depends on the form).

The reproduction countdown is initialized with preoviposition so that when the agent's reproduction is enabled, it automatically enters the preoviposition.

After reproducing, the mother agent $i$ consume energy and get a new reproduction count down:

\[\begin{align} t_{i}^{\left(r\right)} &\gets t^{\left(r\right)}\left(z_{i}^{\left(f\right)}\right),\\ e_{i} &\gets e_{i}-E_{C}. \end{align}\]

See also: get_preoviposition_countdown, get_reproduction_countdown.

RiceBPH.Models.agent_action_die! — Function

agent_action_die!(agent, model)

Check if the agent should be eliminated from the simulation.

The agent $i$ dies and gets removed from the simulated if either: its energy is zero or less ($e_i \le 0$), or its stage is "Dead" ($z^{(s)}_i = 3$).

See also: Stage.

Reproduction parameters

RiceBPH.Models.Form — Type

The form of BPH agent. This enum is Bool based.

Value	Bool	Description

Brachy	false	The truncate-winged form

Macro	true	The fully-winged form

RiceBPH.Models.get_reproduction_countdown — Function

get_reproduction_countdown(form::Form)

Return the reproduction countdown distribution depends on the agents' form.

Macropterous's reproduction countdown distribution is $\mathcal{N}(22.53, 9.28)$.

Brachypterous's reproduction countdown distribution is $\mathcal{N}(15.26, 6.29)$.

RiceBPH.Models.get_preoviposition_countdown — Function

get_preoviposition_countdown(form::Form)

Return the first-reproduction countdown distribution depends on the agents' form.

Macropterous's preoviposition distribution is $\mathcal{N}(177.60, 72.00)$.

Brachypterous's preoviposition distribution is $\mathcal{N}(91.20, 72.00)$.

Age structure

RiceBPH.Models.Stage — Type

Stages of BPHs agent, represented by a Int8 value. The stages are:

Stage	Value	Represents

Egg	0	BPH's eggs

Nymph	1	BPHs in nymph stage

Adult	2	BPHs in adult stage

Dead	3	BPHs in adult stage but their time runs out

Agents with Dead stage will be removed at the end of their step. See also: get_next_stage

RiceBPH.Models.get_stage_countdown — Function

get_stage_countdown(stage::Stage, gender::Gender, form::Form)

Return the count down to the next stage of stage, depends on the gender and form of the agent.

Eggs' stage countdown distribution: $\mathcal{N}(228.00, 84.00)$.

Female nymphs' stage countdown distribution: $\mathcal{N}(324.00, 60.00)$.

Male nymphs' stage countdown distribution: $\mathcal{N}(336.00, 24.00)$.

Male adults' stage countdown distribution: $\mathcal{N}(276.00, 12.00)$.

Female macropterous adults' stage countdown distribution: $\mathcal{N}(660.00, 12.00)$.

Female brachypterous adults' stage countdown distribution: $\mathcal{N}(540.00, 12.00)$.

Dead stage is just a pending removal at the end of the simulation step, therefore the countdown distribution is just the set $\{9999\}$.

RiceBPH.Models.get_next_stage — Function

get_next_stage(stage::Stage)

Return the next age structure of stage. See also Stage.

The stage after Egg is Nymph.

The stage after Nymph is Adult.

The stage after Adult is Dead.

Data collection

Danger

This is unfinished documentation.

RiceBPH.Models.model_action_summarize! — Function

model_action_summarize!(model)

Collect agent and rice statistics and save them in model properties. The collected is the percentage of healthy rice $r_R$ and the number of BPHs, matching stages, forms and genders.

Metric	Description

$r_R$	Percentage of healthy rice

$n_E$	Number of eggs

$n_N$	Number of nymphs

$n_M$	Number of adults with fully-winged form

$n_B$	Number of adults with truncate-winged form

$n_B$	Number of females

These metrics are calculated as follow:

\[\begin{align} r_{R} & =\frac{\left|\left\{ \left(x,y\right)\colon t_{x,y}=1\land e_{x,y}\ge0.5\right\} \right|}{\left|\left\{ \left(x,y\right)\colon t_{x,y}=1\right\} \right|}\\ n_{E} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=0\right\} \right|,\\ n_{N} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=1\right\} \right|,\\ n_{M} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=2\land z_{i}^{\left(t\right)}=0\right\} \right|,\\ n_{B} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=2\land z_{i}^{\left(t\right)}=1\right\} \right|,\\ n_{F} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}\ne0\land z_{i}^{\left(g\right)}=1\right\} \right|. \end{align}\]

Bibliography

Syahrawati2019

M. Syahrawati, O.A. Putra, R. Rusli, E. Sulyanti, Population structure of brown planthopper (Nilaparvata lugens, Hemiptera: Delphacidae) and attack level in endemic area of Padang city, Indonesia, Asian Journal of Agriculture and Biology, 7(Special Issue), 271-276, 2019.