API Reference

Model module

RiceBPH.Models.agent_action_die! — Method

agent_action_die!(agent, model)

Check if the agent should be eliminated from the simulation.

The agent $i$ dies and gets removed from the simulated if either: its energy is zero or less ($e_i \le 0$), or its stage is "Dead" ($z^{(s)}_i = 3$).

See also: Stage.

RiceBPH.Models.agent_action_eat! — Method

agent_action_eat!(agent, model)

Perform eat action of agent.

First, check if the current cell is a FlowerCell, if it is, stop the action.

An amount of energy is subtracted from the RiceCell $x_i,y_i$ and added to the agent $i$.

\[\begin{align} e_{x_i, y_i} &\gets e_{x_i,y_i} - \Delta e,\\ e_i &\gets e_i + \Delta e, \end{align}\]

where the transfered energy $\Delta e$ is calculated by

\[\begin{equation} \Delta e=\min\left(1-e_i,e_{x_i,y_i},e_{T}\right). \end{equation}\]

RiceBPH.Models.agent_action_growup! — Method

agent_action_growup!(agent, model)

Perform the grow up action on agent with id $i$:

The stage countdown of the agent is decresed by one.

\[\begin{equation} t_i^{(s)} \gets t_i^{(s)} - 1. \end{equation}\]

The agent then consumes energy.

\[e_i \gets e_i - E_C\]

If stage countdown is greater than zero, end the action. Otherwise, get the next stage and sample the next stage countdown, assign the next stage and the next countdown to the agent. The next stage $z^{(s)}_i$ and next the stage countdown $t^{(s)}_i$ is returned by get_next_stage.

RiceBPH.Models.agent_action_move! — Method

agent_action_move!(agent, model)

Perform the move action of the agent:

Let the agent's ID $i$, the agent's position on the grid $x=x_i$, $y=y_i$. The agent only moves under certain conditions, which are: energy is greater or equal to energy consumption

\[\begin{equation} e_i \ge E_T; \end{equation}\]

or the cell that the agent is currently on is a rice cell and has zero energy:

\[\begin{equation} e_{x, y} = 0; \end{equation}\]

or the cell type is flower:

\[\begin{equation} t_{x, y} = 0. \end{equation}\]

If the conditions are not met, stop the action. Otherwise, the agent then consumes energy:

\[\begin{equation} e_{i} \gets e_{i} - E_C. \end{equation}\]

After that, a direction $\Delta x, \Delta y$ is sampled within the radius of 2 cells (approximately 30cm) with weights. The weight of the moving direction is calculated by

\[\begin{equation} w_{\Delta x,\Delta y}=\begin{cases} e_{x+\Delta x,y+\Delta y}, & t_{x+\Delta x,y+\Delta y}=1,\\ 0.5, & t_{x+\Delta x,y+\Delta y}=0. \end{cases} \end{equation}\]

After a direction is sampled, the agent then moves along the sampled direction

\[\begin{align} x_i &\gets x_i + \Delta x,\\ y_i &\gets y_i + \Delta y. \end{align}\]

RiceBPH.Models.agent_action_reproduce! — Method

agent_action_reproduce!(agent, model)

Perform the reproductive action of agent $i$.

First, decrease the reproduction countdown of the agent $i$.

\[\begin{equation} t^{(p)}_i \gets t^{(p)}_i - 1 \end{equation}\]

After that, check for the reproduction conditions: the agent is a female,

\[\begin{equation} z^{(g)}_i = 1, \end{equation}\]

the agent energy is larger or equals than the energy consumption

\[\begin{equation} e_i \ge E_C, \end{equation}\]

the reproduction countdown is zero:

\[t^{(p)} = 0,\]

the energy of rice cell at agent position is greater than the energy transfer parameter:

\[e_{x_i, y_i} \ge E_T.\]

If the condition is met, the number of offsprings $N$ from the distributions of offspring quantity $\mathcal{N}(8.50, 3.50)$.

For $k=\overline{1, N}$, the offspring agent $k$ is initialized with:

energy the same as their parent: $e_k \gets e_i$,

the position of the parent $x_k \gets x_i$, $y_k \gets y_i$,

stage is egg: $z^{(s)}_k \gets 0$,

stage countdown is of egg: $t^{(s)}_k \gets$$\mathcal{N}(228.00, 84.00)$,

gender $z^{(g)}_k$: sample from {Male, Female} with weight {0.37, 0.63},

form $z^{(f)}_k$: sample from {Brachy, Macro} with weight {0.47, 0.53},

reproduction countdown follows the preoviposition distribution (which depends on the form).

The reproduction countdown is initialized with preoviposition so that when the agent's reproduction is enabled, it automatically enters the preoviposition.

After reproducing, the mother agent $i$ consume energy and get a new reproduction count down:

\[\begin{align} t_{i}^{\left(r\right)} &\gets t^{\left(r\right)}\left(z_{i}^{\left(f\right)}\right),\\ e_{i} &\gets e_{i}-E_{C}. \end{align}\]

See also: get_preoviposition_countdown, get_reproduction_countdown.

RiceBPH.Models.agent_step! — Method

agent_step!(agent, model)

Perform agent actions in step, actions are determined by their stages. Each stage's actions is defined in its respective step function.

eggs: agent_step_egg!

nymphs: agent_step_nymph!

adults: agent_step_adult!

RiceBPH.Models.agent_step_adult! — Method

agent_step_adult!(agent, model)

The adult BPHs have the following actions (performed in order):

agent_action_growup!

agent_action_move!

agent_action_eat!

agent_action_reproduce!

agent_action_die!

RiceBPH.Models.agent_step_egg! — Method

agent_step_egg!(agent, model)

The eggs have the following actions (performed in order):

agent_action_growup!

agent_action_die!

RiceBPH.Models.agent_step_nymph! — Method

agent_step_nymph!(agent, model)

The nymphs have the following actions (performed in order):

agent_action_growup!

agent_action_move!

agent_action_eat!

agent_action_die!

RiceBPH.Models.get_next_stage — Method

get_next_stage(stage::Stage)

Return the next age structure of stage. See also Stage.

RiceBPH.Models.get_next_stage — Method

The stage after Adult is Dead.

RiceBPH.Models.get_next_stage — Method

The stage after Egg is Nymph.

RiceBPH.Models.get_next_stage — Method

The stage after Nymph is Adult.

RiceBPH.Models.get_preoviposition_countdown — Method

get_preoviposition_countdown(form::Form)

Return the first-reproduction countdown distribution depends on the agents' form.

RiceBPH.Models.get_preoviposition_countdown — Method

Brachypterous's preoviposition distribution is $\mathcal{N}(91.20, 72.00)$.

RiceBPH.Models.get_preoviposition_countdown — Method

Macropterous's preoviposition distribution is $\mathcal{N}(177.60, 72.00)$.

RiceBPH.Models.get_reproduction_countdown — Method

get_reproduction_countdown(form::Form)

Return the reproduction countdown distribution depends on the agents' form.

RiceBPH.Models.get_reproduction_countdown — Method

Brachypterous's reproduction countdown distribution is $\mathcal{N}(15.26, 6.29)$.

RiceBPH.Models.get_reproduction_countdown — Method

Macropterous's reproduction countdown distribution is $\mathcal{N}(22.53, 9.28)$.

RiceBPH.Models.get_stage_countdown — Method

get_stage_countdown(stage::Stage, gender::Gender, form::Form)

Return the count down to the next stage of stage, depends on the gender and form of the agent.

RiceBPH.Models.get_stage_countdown — Method

Female brachypterous adults' stage countdown distribution: $\mathcal{N}(540.00, 12.00)$.

RiceBPH.Models.get_stage_countdown — Method

Female macropterous adults' stage countdown distribution: $\mathcal{N}(660.00, 12.00)$.

RiceBPH.Models.get_stage_countdown — Method

Male adults' stage countdown distribution: $\mathcal{N}(276.00, 12.00)$.

RiceBPH.Models.get_stage_countdown — Method

Dead stage is just a pending removal at the end of the simulation step, therefore the countdown distribution is just the set $\{9999\}$.

RiceBPH.Models.get_stage_countdown — Method

Eggs' stage countdown distribution: $\mathcal{N}(228.00, 84.00)$.

RiceBPH.Models.get_stage_countdown — Method

Female nymphs' stage countdown distribution: $\mathcal{N}(324.00, 60.00)$.

RiceBPH.Models.get_stage_countdown — Method

Male nymphs' stage countdown distribution: $\mathcal{N}(336.00, 24.00)$.

RiceBPH.Models.init_bphs! — Method

init_bphs!(model)

Create $N_I$ agents and put them in the model. The state variables of agent $i$ are initialized as follows:

positions $x_i, y_i$: sampled from cumulative distribution $\mathrm{Poi}(6.00)$ with constraint $x_i \le 15$ and $y_i \le 15$ (see IP_MAX and IP_DST),

energy $e_i$: sampler from $\mathcal{N}(0.50, 0.50)$,

gender $z^{(g)}_i$: sample from {Male, Female} with weight {0.37, 0.63},

form $z^{(f)}_i$: sample from {Brachy, Macro} with weight {0.47, 0.53},

stage $z^{(s)}_i$: sample from {Egg, Nymph, Adult, Dead} with weight {0.33, 0.47, 0.20, 0.00},

stage countdown $t^{(s)}_i$: derives from their other state variables,

reproduction countdown $t^{(r)}_i$: is the preoviposition countdown if agent is not an adult, else the countdown is either preoviposition countdown or reproduction countdown with equal chances.

RiceBPH.Models.init_cell_types — Method

init_cell_types(map_size::Integer, flower_width::Integer)
init_cell_types(S::Integer, S_F::Integer)

Returns a CellType matrix. Rice cells' values are 1. Flower cells' values are 0. The resulting matrix can be defined by:

\[\begin{equation} \left[t_{x,y}\right]_{S \times S}=\left\{ \begin{array}{cl} 0, & y_{a}\le y\le y_{b},\\ 1, & \text{otherwise}, \end{array}\right\}, \end{equation}\]

where $y_a$ and $y_b$ are auxilary variables defined by:

\[\begin{align} y_{a} & =\left\lfloor \frac{S-S_{F}}{2}\right\rfloor +\left(S+S_{F}\mod2\right)+1.\\ y_{b} & =S+S_{F}-1. \end{align}\]

See also: CellType.

RiceBPH.Models.init_model — Method

init_model(; seed::Union{Int, Nothing} = Nothing, kwargs...)

Create and return a ABM model. Keyword arguments are passed to ModelParameters.

At the initialization of the model, inferable model properties are calculated to optimize the runtime of the simulation (see init_properties). This process also create the simulated environment and the vegetation cell state.

After that, the BPH agents are initialized with random states. After the BPHs are initialized and placed inside the model, we collect the initial data of interest and store them in the model state.

RiceBPH.Models.init_positions — Method

init_positions(rng, position::InitPosition, num_bphs::Integer, map_size::Integer)

Return an iterator over all the initial positions.

See also: IP_PTS, IP_DST, InitPosition.

RiceBPH.Models.init_pr_eliminate — Method

init_pr_eliminate(init_pr::Float32, cell_types::Matrix{CellType})

Returns Float32 matrix of elimination probability. The result matrix is defined by:

\[\left[p_{x,y}\right]_{S\times S} = P_0 \times (1 - M)) ∗ G(\sigma),\]

where:

$G(\sigma)$ is the Gaussian kernel,

$P_0$ is the base elimination probability,

$M$ is the cell type matrix (see init_cell_types).

RiceBPH.Models.init_properties — Method

init_properties(parameters::ModelParameters)

Construct a ModelProperties from ModelParameters.

RiceBPH.Models.init_rice_map — Method

init_rice_map(S::Integer)

Returns a matrix filled with $1$ of size $S \times S$.

\[\left[e_{x,y}\right]_{S\times S}\equiv1.\]

In code terms, it's basically ones(Float32, S, S).

RiceBPH.Models.model_action_eliminate! — Method

model_action_eliminate!(model)

Eliminate BPHs base on their energy and the pr eliminate map. The elimination probability is given by $p_{x,y}$.

For each position $x, y$, for each agent $i$ whose position $x_i = x, y_i = y$, the agent is removed from the simulation with the probability $p_{x, y}$.

RiceBPH.Models.model_action_summarize! — Method

model_action_summarize!(model)

Collect agent and rice statistics and save them in model properties. The collected is the percentage of healthy rice $r_R$ and the number of BPHs, matching stages, forms and genders.

Metric	Description

$r_R$	Percentage of healthy rice

$n_E$	Number of eggs

$n_N$	Number of nymphs

$n_M$	Number of adults with fully-winged form

$n_B$	Number of adults with truncate-winged form

$n_B$	Number of females

These metrics are calculated as follow:

\[\begin{align} r_{R} & =\frac{\left|\left\{ \left(x,y\right)\colon t_{x,y}=1\land e_{x,y}\ge0.5\right\} \right|}{\left|\left\{ \left(x,y\right)\colon t_{x,y}=1\right\} \right|}\\ n_{E} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=0\right\} \right|,\\ n_{N} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=1\right\} \right|,\\ n_{M} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=2\land z_{i}^{\left(t\right)}=0\right\} \right|,\\ n_{B} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}=2\land z_{i}^{\left(t\right)}=1\right\} \right|,\\ n_{F} & =\left|\left\{ i\colon z_{i}^{\left(s\right)}\ne0\land z_{i}^{\left(g\right)}=1\right\} \right|. \end{align}\]

RiceBPH.Models.model_step! — Method

model_step!(model)

Model behaviour in one step, which includes:

model_action_eliminate!

model_action_summarize!

RiceBPH.Models.normal_hour_range — Method

normal_hour_range(a, b)

Return normal_range(a, b; mul=24). See normal_range for details.

RiceBPH.Models.normal_range — Method

normal_range(a, b; mul=1)

Return a Normal distribution with mean $\mu'$ and standard deviation $\sigma'$ determined by:

\[\begin{align} a & =\mu-\sigma,\\ b & =\mu+\sigma. \end{align}\]

The found mean and standard deviation is scaled by mul variable to obtain the final mean and deviation.

\[\begin{align} \mu' &= \mu \cdot \mathrm{mul}, \\ \sigma' &= \sigma \cdot \mathrm{mul}. \end{align}\]

RiceBPH.Models.normalize — Method

normalize(v)

Return a probability vector $v'$ with $v$ as the Weight:

\[v' = \frac{v}{\sum_{i} v_i}.\]

RiceBPH.Models.show_dist — Method

show_dist(d)

Return LaTeX string for normal distribution.

RiceBPH.Models.show_dist — Method

show_dist(d)

Return LaTeX string for poisson distribution.

RiceBPH.Models.show_dist — Method

show_dist(w)

Return LaTeX string for weight vectors.

RiceBPH.Models.show_enum — Method

show_enum(enum)

Return string format for enum

RiceBPH.Models.BPH — Type

The BPH agents, with the following state variables.

Variable	Type	Description

id	Int	The agent identifier number

pos	Dims{2}	The agent's position on the environment grid

energy	Float16	The agent's energy

gender	Gender	The agent's gender

form	Form	The agent's form

stage	Stage	The agent's current stage

stage_cd	Int16	The countdown to agent's next stage

reproduction_cd	Int16	The countdown to agent's next reproduction

RiceBPH.Models.CellType — Type

The enum CellType is based on Bool values, indicates a cell is a rice cell or flower cell:

Value	Bool	Int

FlowerCell	false	0

RiceCell	true	1

RiceBPH.Models.Gender — Type

The gender of BPH agent. This enum is Bool based.

Value	Int	Bool

Male	0	false

Female	1	true

RiceBPH.Models.InitPosition — Type

BPH Initialization position.

Value	Description

Corner	1meters at a corner of the map

Border	1 meters along a border of the map, the border does not overlap with flower

See also IP_DST.

RiceBPH.Models.ModelParameters — Type

ModelParameters contains initialization variables.

Variable	Type	Default	Description	Symbol

map_size	Int	-	The environment grid size	$S$

flower_width	Int	-	The number of flower cells at the center of the grid	$S_F$

init_num_bphs	Int	-	The number of initialized BPH	$N_{I}$

init_pr_eliminate	Float32	-	The base elimination probability	$P_0$

init_position	InitPosition	Corner	The BPHs' initial positions	-

energy_transfer	Float32	0.032	The energy conversion from rice to BPHs.	$E_T$

RiceBPH.Models.ModelProperties — Type

Model properties. There are three types of properties:

Model state

The first types are properties that are inferred from the input variables to be used during the simulation.

The second types are properties to collect statistics during the simulations.

Properties (except num_rice_cells) those are used to collect model statistics has the name prefix .

Model states and inferable parameters:

Name	Type	Description	Symbol

parameters	ModelParameters	The initialisation parameter	-

rice_map	Matrix{Float32}	The matrix $[e_{x,y}]_{s\times s}$ of rice cells' energy	-

pr_eliminate_map	Matrix{Float32}	The matrix of elimination probabilities at each grid position	-

cell_types	Matrix{CellType}	The matrix contains the cell type of each grid position	-

eliminate_positions	Vector{CartesianIndex}	Grid positions with elimination probability $p_{x,y} > 0$	-

rice_positions	Vector{CartesianIndex}	The positions of RiceCell.	-

num_rice_cells	Int	Total number of rice cells, equals to $S^2$	-

energy_consume	Float32	The energy consumption $E_C$ of BPH agents	$E_C$

moving_directions	Vector{Tuple{Int, Int}}	All the possible moving directions	-

The energy consumption $E_C$ is calculated by dividing the energy transfer $E_T$ to all the energy-consuming actions (agent_action_move!, agent_action_reproduce!, agent_action_growup!):

\[\begin{equation} E_C = \frac{E_T}{\mathrm{number of actions}} = \frac{E_T}{3}. \begin{equation}\]

Statistics

These data are collected at each time step, see also model_action_summarize!.

Name	Type	Description	Symbol

pct_rices	Float32	Percentage of healty rice	$r_{R}$

num_eggs	Int	Number of BPH eggs	$n_{E}$

num_nymphs	Int	Number of BPH in nymphs stage	$n_{N}$

num_brachys	Int	Number of BPH in adult stage, truncate-winged form	$n_{B}$

num_macros	Int	Number of BPH in adult stage, fully-winged form	$n_{M}$

num_females	Int	Number of female BPHs, counting nymphs and adults	$n_{F}$

Results module

Base.isvalid — Method

Base.isvalid(x::Real)

The isvalid function checks whether a given floating-point number x is valid, i.e., it is neither NaN nor infinite.

RiceBPH.Results.dropnan — Method

dropnan(df)

Remove rows with NaN from dataframe, returns a subdataframe.

RiceBPH.Results.fit_fn — Method

RiceBPH.Results.get_data_names — Method

get_data_names(df::DataFrame)

Return a vector of the names of the data columns.

RiceBPH.Results.get_factor_name — Method

get_factor_name(df::DataFrame)::Symbol

Return the variable factor from the OFAAT result.

RiceBPH.Results.get_stable_bph_timesteps — Method

get_stable_bph_timesteps(num_bphs)

Return a range of timesteps where the population of BPHs is stable. This time period is defined as a week before the BPH population peak.

RiceBPH.Results.get_stats — Method

get_stats(df)

Get statistics over all replications of each parameter from the OFAAT result.

RiceBPH.Results.infer_stats! — Method

infer_stats!(df:DataFrame)

Calculate inferable statistics from the result dataframe. Return the data frame. This function is inplace. The inferable statistics are:

num_bphs: total number of BPHs

pct_nymphs: percentage of nymphs

pct_macros: percentage of macros

pct_bracys: percentage of bracys

pct_females percentage of females BPH

pct_males: percentage of males BPH

Eggs are not counted in the calculations.

RiceBPH.Results.load — Method

load(path_to_jdf_folder)

Load the result, apply some inference on the results (to get extra statistics). Return the a dataframe of the results.

RiceBPH.Results.mask_nonconverge — Method

masknonconverge(fitdf)

Change the row which is not converged to NaN.

RiceBPH.Results.qcv — Method

function qcv(x)

The qcv function calculates the coefficient of dispersion (CD) for a given dataset x using the interquartile range (IQR).

Parameters

x: An array or collection of numerical values representing the dataset for which the CD is to be calculated.

Returns

The function returns a single floating-point value representing the coefficient of dispersion (CD) of the dataset.

Description

The coefficient of dispersion (CD) is a measure of the relative variability or dispersion of a dataset. It is calculated as the absolute difference between the first quartile (Q1) and the third quartile (Q3), divided by their sum.

The qcv function calculates the CD by first computing the first quartile q1 and the third quartile q3 using the quantile function from the Statistics module. The quantile function returns the value that corresponds to the given quantile of the dataset. In this case, q1 represents the 25th percentile (Q1), and q3 represents the 75th percentile (Q3).

The CD is then computed as the absolute difference between q3 and q1, divided by their sum.

Example Usage

data = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
cd = qcv(data)
println("Coefficient of Dispersion: ", cd)

Note

The quantile function used in the qcv function assumes the default values for the beta and alpha parameters, which correspond to the definitions of the quartiles used in traditional statistics. However, these parameters can be adjusted if desired.

The qcv function assumes that the input dataset x is numeric. Providing non-numeric values may result in unexpected behavior or errors.

RiceBPH.OFAAT.run_ofaat! — Method

run_ofaat!(num_steps, num_reps, factor, values; kw...)

Run one-factor-at-a-time replication experiment. Return the combined dataframe created by concatnating the result dataframes from all the replication. This function is the same as paramscan from Agents.jl, but paramscan does not do garbage collection, this one does to prevent out of memory error.

RiceBPH.OFAAT.run_replicate! — Method

run_replicate!(num_steps, seed, factor, value, kw)

Run a replication for num_steps steps. The other parameters are used to initialize the model. Returns model dataframe. The final dataframe contains extra columns to represent seed and factor.

Visualisations

RiceBPH.Visualisations.FLOWER_COLOR — Constant

Color of the flower cells on the heatmap (#FFC107).

RiceBPH.Visualisations.RICE_COLORS — Constant

Color of the rice cells on the heatmap, the color is RGB(0.0, 0.8, 0.0, energy) where energy is the energy of the rice cell.

Constants

Population structure related

Reproduction related

RiceBPH.Models.AVG_EGGS_M — Constant

Average total number of eggs (fully winged form)

RiceBPH.Models.AVG_EGGS_B — Constant

Average total number of eggs (truncate winged form)